+++ custom_components/homeseer_bridge/const.py
@@ -6,7 +6,7 @@ CONF_EXCLUDED_TERMS = "excluded_terms"
 
 DEFAULT_HS_URL = "http://192.168.0.193"
 DEFAULT_MQTT_PREFIX = "Homeseer/Chip23/mcsMQTT"
-DEFAULT_EXCLUDED_TERMS = "august,yolink,shelly"
+DEFAULT_EXCLUDED_TERMS = "august,yolink,shelly,ultraweatherwu3,narrative"
 
 PLATFORMS = ["switch", "light", "sensor", "binary_sensor", "lock", "cover", "fan"]